Added ember-cli-github-pages addon
