+++ app/controllers/events/view/settings/options.js
@@ -1,7 +1,7 @@
 import Controller from '@ember/controller';
 import { action } from '@ember/object';
 
-export default class extends Controller {
+export default class OptionsController extends Controller {
 
   @action
   openDeleteEventModal() {
