+++ app/routes/events/view/settings.js
@@ -6,20 +6,4 @@ export default class SettingsRoute extends Route {
   titleToken() {
     return this.l10n.t('Settings');
   }
-
-  beforeModel() {
-    const { currentUser } = this.authManager;
-    if (!(currentUser.isAnAdmin || this.modelFor('events.view').owner.get('email') === currentUser.email)) {
-      this.transitionTo('events.view');
-    }
-  }
-
-  async model() {
-    const eventDetails = this.modelFor('events.view');
-    return {
-      event       : await eventDetails,
-      roleInvites : await eventDetails.query('roleInvites', {}),
-      roles       : await this.store.findAll('role')
-    };
-  }
 }
