+++ app/styles/components/all.scss
@@ -2,7 +2,7 @@
 @import 'notification-dropdown';
 @import 'speaker-list';
 @import 'explore/side-bar';
-@import 'events/view/export';
+@import 'events/view/settings/export';
 @import 'tabbed-navigation';
 @import 'ui-table';
 @import 'scheduler';
