+++ app/templates/events/view.hbs
@@ -1,4 +1,4 @@
-{{#if (and (or (eq this.session.currentRouteName 'events.view.index') (eq this.session.currentRouteName 'events.view.export')) (gt this.unverified-user-message.length 0))}} 
+{{#if (and (or (eq this.session.currentRouteName 'events.view.index') (eq this.session.currentRouteName 'events.view.settings.export')) (gt this.unverified-user-message.length 0))}} 
   {{#if (eq this.model.state 'draft')}}
     <UnverifiedUserMessage />
   {{/if}}
@@ -42,9 +42,6 @@
         </LinkTo>
         <LinkTo @route="events.view.videoroom" class="item">
           {{t 'Video'}}
-        </LinkTo>
-        <LinkTo @route="events.view.export" class="item">
-          {{t 'Export'}}
         </LinkTo>
         {{#if (or this.authManager.currentUser.isAnAdmin (eq this.model.owner.email this.authManager.currentUser.email))}}
           <LinkTo @route="events.view.settings" class="item">
