+++ app/templates/events/view/settings/options.hbs
@@ -1,0 +1,53 @@
+<div class="row">
+  <div class="ui red segment">
+    <div class="ui grid">
+      <div class="ui left aligned sixteen wide mobile eight wide computer column">
+        <h3 class="ui header">
+          {{t 'Transfer Ownership'}}
+        </h3>
+        <p class="ui muted text">
+          {{t 'You\'ll lose all the owner rights once the other user accepts the invitation for ownership.'}}
+        </p>
+      </div>
+      <div class="ui right aligned sixteen wide mobile eight wide computer column">
+        <button {{action 'openEventTransferModal' this.model.event.id this.model.event.name}} class='ui labeled icon red button'>
+          <i class="share icon"></i>
+          {{t 'Transfer Event'}}
+        </button>
+      </div>
+
+      <div class="ui left aligned sixteen wide mobile eight wide computer column">
+        <h3 class="ui header">
+          {{t 'Delete this event'}}
+        </h3>
+        <div class="ui muted text">
+          {{t 'Once you delete an event, you will lose all the data related to this event.'}}
+        </div>
+      </div>
+      <div class="ui right aligned sixteen wide mobile eight wide computer column">
+        <button {{action 'openDeleteEventModal'}} class='ui labeled icon red button'>
+          <i class="trash icon"></i>
+          {{t 'Delete Event'}}
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
+<Modals::EventTransferModal
+  @isLoading={{this.isLoading}}
+  @isOpen={{this.isEventTransferModalOpen}}
+  @confirmEventName={{this.confirmEventName}}
+  @eventName={{this.eventName}}
+  @openConfirmEventTransferModal={{action "openConfirmEventTransferModal"}}
+  @transferEvent={{action "transferEvent" this.model}} />
+<Modals::ConfirmEventTransferModal
+  @currentInvite={{this.currentInvite}}
+  @isLoading={{this.isLoading}}
+  @isOpen={{this.isConfirmEventTransferModalOpen}}
+  @checked={{this.checked}}
+  @transferEvent={{action "transferEvent"}} />
+<Modals::EventDeleteModal
+  @isLoading={{this.isLoading}}
+  @isOpen={{this.isEventDeleteModalOpen}}
+  @eventName={{this.model.event.name}}
+  @deleteEvent={{action "deleteEvent"}} />
