+++ package.json
@@ -81,6 +81,7 @@
     "ember-cli-dotenv": "^3.1.0",
     "ember-cli-fastboot": "^2.2.3",
     "ember-cli-flash": "^2.0.0",
+    "ember-cli-github-pages": "^0.2.2",
     "ember-cli-head": "^1.0.0",
     "ember-cli-html-minifier": "^1.1.0",
     "ember-cli-htmlbars": "^5.3.1",
